requirements.txt
cffi==2.0.0
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pycparser==2.23
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
